+++ src/types/async.ts
@@ -1,0 +1,4 @@
+/**
+ * Type that represents pure value or value obtained by promise
+ */
+export type PromiseOr<TValue> = Promise<TValue>|TValue;