+++ src/types/pageable.ts
@@ -1,0 +1,15 @@
+/**
+ * Interface for paging
+ */
+export interface Pageable
+{
+    /**
+     * Number of items that should be returned
+     */
+    size: number;
+
+    /**
+     * Page number that should be returned
+     */
+    page: number;
+}